Refactored my code to reduce repeated code: -3 controllers --> 1 controller -3 inputDatas --> 2 inputDatas -3 inputBoundaries --> 1 inputBoundary -3 use case interactors --> 1 use case interactor -3 outputBoundaries --> 1 outputBoudnary -3 presenters --> 1 presenter
This should adhere better to the standards of clean architecture and general coding principles.

I have also implemented functioning tests for each method of the new use case interactor.

Feel free to ask any questions about this code, as I hope it may be of help if anyone wants it.

(Currently the old implementations are commented out but this is temporary for this push, as I want the old system to be uploaded in case the new one fails at some point. I will remove it immediately after in the next push.)